fix desktop view, and memoize
